YT-agent/requirements.txt
claude-agent-sdk==0.1.50
python-dotenv==1.1.0
